experimental-simulation-ms/requirements.txt
# requires Python >= 3.12.3
astropy==7.1.0
cloudpickle==3.1.1
ipykernel==6.30.1
mammos-entity==0.9.0
mammos-units==0.3.1.post0
matplotlib==3.10.3
numpy==2.2.6
pandas==2.2.3
pymatgen==2025.6.14
pysr==1.5.9
PyYAML==6.0.2
scikit-learn==1.7.0
scipy==1.16.1
seaborn==0.13.2
statsmodels==0.14.5
sympy==1.14.0
tabulate==0.9.0
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
tqdm==4.67.1
umap-learn==0.5.9.post2
xgboost==3.0.3